+++ Neighbors/Services/DAL/CategoriesRepository.cs
@@ -30,14 +30,17 @@ namespace Neighbors.Services.DAL
 
         public async Task<int> DeleteCategory(int categoryId)
         {
-            var category = await _context.Categories.FindAsync(categoryId);
+            var category = await GetCategoryById(categoryId);
             if (category != null)
             {
-                _context.Categories.Remove(category);
+                if (category.Products.Count == 0)
+                {
+                    _context.Categories.Remove(category);
+                } else {
+                    return -1;
+                }
             }
-
-            return await _context.SaveChangesAsync();
-           
+            return await _context.SaveChangesAsync(); 
         }
 
         public async Task<int> UpdateCategory(int categoryId, Category category)
@@ -51,11 +54,15 @@ namespace Neighbors.Services.DAL
         #region Getters
         public async Task<ICollection<Category>> GetAllCategories()
         {
-            //return await _context.Categories.Include(p => p.Products).ToListAsync();
-            return await _context.Categories.ToListAsync();
+            return await _context.Categories.Include(p => p.Products).ToListAsync();
         }
 
-        public async Task<Category> GetCategoryById(int id)
+		public async Task<ICollection<Category>> GetAllCategoriesShort()
+		{
+			return await _context.Categories.ToListAsync();
+		}
+
+		public async Task<Category> GetCategoryById(int id)
         {
             return (await _context.Categories.Include(p => p.Products).FirstOrDefaultAsync(c => c.Id == id)); 
         }
